Extract component OverviewTab
