+++ frontend/src/scenes/instance/SystemStatus/index.tsx
@@ -1,40 +1,17 @@
 import './index.scss'
 
 import React from 'react'
-import { Alert, Table, Tag, Card } from 'antd'
+import { Alert } from 'antd'
 import { systemStatusLogic } from './systemStatusLogic'
 import { useValues } from 'kea'
 import { PageHeader } from 'lib/components/PageHeader'
-import { SystemStatusSubrows } from '~/types'
 import { preflightLogic } from 'scenes/PreflightCheck/logic'
 import { IconExternalLink } from 'lib/components/icons'
-
-function RenderValue(value: any): JSX.Element | string {
-    if (typeof value === 'boolean') {
-        return <Tag color={value ? 'success' : 'error'}>{value ? 'Yes' : 'No'}</Tag>
-    }
-    if (value === null || value === undefined || value === '') {
-        return <Tag>Unknown</Tag>
-    }
-    return value.toString()
-}
+import { OverviewTab } from 'scenes/instance/SystemStatus/OverviewTab'
 
 export function SystemStatus(): JSX.Element {
-    const { systemStatus, systemStatusLoading, error } = useValues(systemStatusLogic)
-    const { configOptions, preflight, preflightLoading, siteUrlMisconfigured } = useValues(preflightLogic)
-
-    const columns = [
-        {
-            title: 'Metric',
-            dataIndex: 'metric',
-            className: 'metric-column',
-        },
-        {
-            title: 'Value',
-            dataIndex: 'value',
-            render: RenderValue,
-        },
-    ]
+    const { error } = useValues(systemStatusLogic)
+    const { preflight, siteUrlMisconfigured } = useValues(preflightLogic)
 
     return (
         <div className="system-status-scene">
@@ -56,8 +33,11 @@ export function SystemStatus(): JSX.Element {
                     description={
                         <>
                             Your <code>SITE_URL</code> environment variable seems misconfigured. Your{' '}
-                            <code>SITE_URL</code> is set to <b>{RenderValue(preflight?.site_url)}</b> but you're
-                            currently browsing this page from{' '}
+                            <code>SITE_URL</code> is set to{' '}
+                            <b>
+                                <code>{preflight?.site_url}</code>
+                            </b>{' '}
+                            but you're currently browsing this page from{' '}
                             <b>
                                 <code>{window.location.origin}</code>
                             </b>
@@ -77,48 +57,7 @@ export function SystemStatus(): JSX.Element {
                     style={{ marginBottom: 32 }}
                 />
             )}
-            <Card>
-                <h3 className="l3">Key metrics</h3>
-                <Table
-                    className="system-status-table"
-                    size="small"
-                    rowKey="metric"
-                    pagination={{ pageSize: 99999, hideOnSinglePage: true }}
-                    dataSource={systemStatus}
-                    columns={columns}
-                    loading={systemStatusLoading}
-                    expandable={{
-                        expandedRowRender: function renderExpand(row) {
-                            return row.subrows ? <Subrows {...row.subrows} /> : null
-                        },
-                        rowExpandable: (row) => !!row.subrows && row.subrows.rows.length > 0,
-                        expandRowByClick: true,
-                    }}
-                />
-            </Card>
-            <Card style={{ marginTop: 32 }}>
-                <h3 className="l3">Configuration options</h3>
-                <Table
-                    className="system-config-table"
-                    size="small"
-                    rowKey="metric"
-                    pagination={{ pageSize: 99999, hideOnSinglePage: true }}
-                    dataSource={configOptions}
-                    columns={columns}
-                    loading={preflightLoading}
-                />
-            </Card>
+            <OverviewTab />
         </div>
-    )
-}
-
-function Subrows(props: SystemStatusSubrows): JSX.Element {
-    return (
-        <Table
-            rowKey="metric"
-            pagination={{ pageSize: 99999, hideOnSinglePage: true }}
-            dataSource={props.rows}
-            columns={props.columns.map((title, dataIndex) => ({ title, dataIndex }))}
-        />
     )
 }
